Added CRAM_OPT_SHARED_REF for mpileup too.
It shares the same caveat regarding requiring matching headers.
Do we need a command line option to disable/enable this?

In an ideal world the refs_t structure would be internally indexed by
the fai file and/or on-disk md5 filenames and not by the Nth @SQ
record, meaning different headers could share the same ref cache, but
that's substantially more work to do.